Update code from merges
Purge unused classes
Add comments to BattleHandler for understanding
Begin merge of networked battle packets with Battle system

+++ Unity/Assets/Scripts/Battles/Editor/BattleTesterEditor.cs
@@ -80,12 +80,11 @@ public class BattleTesterEditor : Editor
 
             if (GUILayout.Button("Other selects random attack"))
             {
-                int selectionIndex = Random.Range(0,
+                uint selectionIndex = (uint)Random.Range(0,
                     t.battleHandler.OtherKeeper.monsters[t.battleHandler.OtherCretinIndex].GetAvailableAttacks
                         .Count);
 
-                t.battleHandler.SendBattleOption(false, GameState.Player.EnumBattleSelection.ATTACK,
-                    selectionIndex);
+                t.battleHandler.SendBattleOption(false, GameState.Player.EnumBattleSelection.ATTACK, selectionIndex);
             }
 
             if (GUILayout.Button("Other sends out random cretin"))
@@ -93,14 +92,13 @@ public class BattleTesterEditor : Editor
                 int whileCount = 0;
                 while (whileCount < 20)
                 {
-                    int selectionIndex = Random.Range(0, t.battleHandler.OtherKeeper.monsters.Count);
+                    uint selectionIndex = (uint)Random.Range(0, t.battleHandler.OtherKeeper.monsters.Count);
 
-                    if (t.battleHandler.OtherKeeper.monsters[selectionIndex].CurrentHP > 0)
+                    if (t.battleHandler.OtherKeeper.monsters[(int)selectionIndex].CurrentHP > 0)
                     {
                         if (selectionIndex != t.battleHandler.OtherCretinIndex)
                         {
-                            t.battleHandler.SendBattleOption(false, GameState.Player.EnumBattleSelection.SWAP,
-                                selectionIndex);
+                            t.battleHandler.SendBattleOption(false, GameState.Player.EnumBattleSelection.SWAP, selectionIndex);
                             break;
                         }
                     }
